requirements.txt
fastapi
uvicorn
pydantic
langgraph
langchain-google-genai
langchain-core
python-dotenv
